Support registering RAM blocks of 4GB and larger
HAX_VM_IOCTL_ALLOC_RAM takes a 32-bit size, whereas QEMU RAM blocks
can be 4GB or larger.

Add a new ioctl that takes a 64-bit size but is otherwise the same
as ALLOC_RAM. Define a new capability flag for this new feature, so
as not to break backward compatibility.

With a corresponding QEMU-side change, it is now possible to launch
guests with >= 4GB of RAM.

Fixes #13.