page_walker: Fix 32-bit GVA-GPA translation bug
On Mac, with certain 32-bit guests (PAE off), vcpu_translate()
sometimes fails or yields an invalid GPA like 0xffffffff801c8000.
The root cause is that Clang somehow uses sign extension when
converting the result of left-shifting a uint32_t bit-field to
uint64_t.

If the bit-field is first converted to uint64_t and then shifted,
the end result will be correct. With this patch, one is able to
boot debian_squeeze_i386_desktop.qcow2 with a sufficiently large
guest RAM size (e.g. 3GB) on Mac.