core: fix bug that INVEPT is not invoked on Mac
There is coexistence issue in MacOS when some Mac VMMs are running
at the same time. invept might not be invoked due to coexistence
issue. This change is to fix the issue that invept is not invoked.

It also refactors the existed logic for VMXON/VMXOFF into two
functions. The function names come from PR #49.