Fix checks for IA32_TSC_AUX MSR availability
The IA32_TSC_AUX MSR is not available on some very old CPUs, e.g.
Core 2 Duo. Commit e9f8c89 wanted to prevent the physical MSR from
being accessed if the host CPU does not have it, but unfortunately
ended up implementing the opposite logic. Moreover, the checks
added by the said commit will be skipped if the host CPU lacks an
unrelated feature (Architectural Performance Monitoring, or APM).

Fix the bad host feature checks, and also add a check to guard the
enabling of the Enable RDTSCP VM-execution control, so that:

a) If the host supports RDTSCP, the guest has its own copy of the
IA32_TSC_AUX MSR (separate from the host MSR), and is allowed to
use the RDTSCP instruction.

b) If the host does not support RDTSCP, HAXM does not access
IA32_TSC_AUX (thereby avoiding a host crash), and does not allow
the guest to use the RDTSCP instruction.

Fixes #101.

Signed-off-by: Yu Ning <yu.ning@intel.com>