Generalized page-related constants
Platform-specific headers, e.g. hax_windows.h, defined page-related constants such as page_size and page_shift that are identical in all systems since they are unique to the x86 architecture.

This commit moves the definitions to hax_types.h, and:
- Makes these macros uppercase (as per coding convention).
- Adds the `HAX_` prefix (as per coding convention). This is also necessary to prevent clashes with kernel headers.

These macros are used in all of `core`. Platform-specific code is still allowed to use platform-specific macros like the ones defined in the WDK.

Signed-off-by: Alexandro Sanchez Bach <asanchez@kryptoslogic.com>