haxm debugging feature improvement
This change is to improve robustness and performance for debug
feature(#81). DR registers are loaded to guest only if they are
updated or current thread is migrated to other CPU core. Host
DRs will not be saved and restored unless host process (QEMU)
is debugged with hardware breakpoint.

When a VCPU is being debugged by the host using HW breakpoint,
prevent its DR registers from being tampered with by the guest
OS or by SET_REGS ioctl.

Verified:
 - Guest gdb only;
 - QEMU gdb (debugging guest from host) only;
 - Starting guest gdb, then using QEMU gdb to debug;
 - Host debugging QEMU + Guest gdb;
 - Host debugging QEMU + QEMU gdb;
 - Host debugging QEMU + QEMU gdb + Guest gdb
   QEMU gdb works, and guest gdb reports app crash. Please refer
   to known issues

Known issue:
 - Guest gdb might report app crash When QEMU gdb is enabled
   This is probabaly because existed guest DR state is discarded
   as soon as QEMU gdb enables hardware breakpoint.

Signed-off-by: Junxiao Chang <junxiao.chang@intel.com>