Refactored VMCS dumping code
The previous implementation of `dump_vmcs` obtained the VMCS component name by hashing its value with a gperf-generated perfect hash function at vmcs_names.c. These compact hashes were used as indices in a manually-initialized table of names. This results in lots of boilerplate code, manual initialization/deinitialization and a dependency in gperf.

This new approach uses a switch/case to dispatch the component names and compactifies the component encoding with the macro `HASH` to encourage the compiler to implement the switch/case via jump tables. The approach works in GCC, Clang, MSVC. Advantages:

- Less code.
- No external dependencies.
- No memory allocated at runtime.
- Better performance: No need to query an extra table to compute the perfect hash.
- Perfect hashing tested at build-time: If the hash macro maps two components to the same value, the file won't compile.

Signed-off-by: Alexandro Sanchez Bach <asanchez@kryptoslogic.com>