Verbose error descriptions in emulator tests
Previously, we added `EXPECT_EQ(memcmp(&vcpu, &vcpu_expected, sizeof(test_cpu_t)), 0);` as a simple check at the end of EmulatorTest::run. While this was enough to catch implementation issues in the emulator, it did specify where the exactly was the mismatch happening. Consequently, we often turn to the debugger and manually diff both `test_cpu_t` structures which is time consuming.

This patch adds a dedicated `EmulatorTest::verify` function that checks each register individually and reports any issues to the stderr, before comparing and asserting the equality of both structures. Note that the previous `EXPECT_EQ` has been replaced with `ASSERT_EQ`, as there is no point in keep trying other instruction variants if there's one broken already.

Signed-off-by: Alexandro Sanchez Bach <asanchez@kryptoslogic.com>