Fixed missing flags in grouped instructions
After finding the opcode structure corresponding to the current instruction, we cache `opcode->flags` into the `flags` variable and perform several checks for particular flags such as `INSN_BYTEOP` and `INSN_STACK`.

The issue here is that for instructions that belong in a group, see `INSN_GROUP`, we need to merge the group flags and the flags of the actual opcode before performing such queries. Consequently we need to reposition all `(flags & INSN_*)` statements after we have dealt with groups.

The only exception here is `(flags & INSN_MODRM)` since we need to fetch the ModRM byte in order to figure out which instruction we are dealing with within that group.

Signed-off-by: Alexandro Sanchez Bach <asanchez@kryptoslogic.com>