Skip some checks for host-initiated EFER writes
Although the main purpose of handle_msr_write() is to handle guest-
initiated MSR writes, it is also used by the SET_MSRS ioctl handler
to deal with host-initiated writes to guest MSRs. In the case of
IA32_EFER, these two scenarios actually require the hypervisor to
perform different checks.

Commit fe8aa45 ("Added EFER value checks for handle_msr_write")
added additional checks on IA32_EFER, but some of them are only
applicable to guest-initiated EFER writes. As a result, Android
Emulator always runs into a "VCPU shutdown request" (HAXM panic)
when booting from an AVD snapshot. That is because it tries to
put the vCPUs directly in 64-bit mode, but one of the new checks
prevents it from loading the saved EFER state. In the end, 0 is
loaded into EFER, and the first VM entry fails due to invalid
guest state (vCPU is not in 64-bit mode but uses a 64-bit RIP).

Fix that by adding a Boolean parameter to handle_msr_write() that
helps distinguish between the two scenarios, and skipping all
guest-specific checks for host-initiated EFER writes.

Signed-off-by: Yu Ning <yu.ning@intel.com>