Fix incorrect interruptibility_state before vmx entry.
According to SDM 26.3.1.5 Checks on Guest Non-Register State, Bit 0
(blocking by STI) must be 0 if the IF flag (bit 9) is 0 in the RFLAGS
field.
There is an issue during snapshot loading, that IF and
interruptibility_state don't pass the checks, which will result in
VM-entry failure due to invalid guest state.
This WA correct the bit so that vmx entry check could pass. The normal
interruptibility_state update is done when advancing the IP.
In future, proper approach is expected to replace the WA.

Signed-off-by: Colin Xu <colin.xu@intel.com>