cpuid: Optimize CPUID storage structure
Use the array of hax_cpuid_entry to store feature set. This approach is
helpful for the extension of more CPUID features support. It ensures
that the same data structure hax_cpuid_entry is used internally and
externally. When adding a new CPUID feature, it makes the implementation
easy by inserting an entry to the array of kCpuidSet.

Signed-off-by: Wenchao Wang <wenchao.wang@intel.com>