Fixed releases url in README
The previous url opened a view with only the single latest release, which might be "CheckTool" or "HAXM", depending on the current release schedule. 

The new url brings up a reverse-chronological view of the latest releases, so the same information is available at the same place, but scrolling down will give you more recent releases. On the previous url, you couldn't find downloads for HAXM right now, only CheckTool.