cpuid: Add CPUID manager
Add CPUID manager to manage the supported guest CPUID instructions. For
the leaves in the manager list, if the execution function is specified,
the return values of the host CPUID instruction will be overwritten; If
not specified, the host values will be returned directly. For the leaves
not in the manager list, it will return that those CPUID instructions
are not supported.

Signed-off-by: Wenchao Wang <wenchao.wang@intel.com>