Reset DR7 before setting DR0
Without resetting DR7, spurious exceptions encountered in unexpected
conditions may lead to privilege escalation.
The exploitability depends a lot on the host system, and how it
processes #DB exceptions. Typically if the #DB handler has certain
conditions that make it execute SWAPGS, then the ISR will execute with a
guest-controlled TLS (the IA32_KERNEL_GS_BASE MSR, which contains a
guest-controlled pointer), and here a privilege escalation is possible.

Signed-off-by: Wenchao Wang <wenchao.wang@intel.com>