Check instruction offset is valid before accessing
Check the bytes being used in functions insn_fetch_*() do not exceed the
15-byte stack buffer where they are stored.
It is easy for the guest to trigger out-of-bounds accesses in the host
by just racing the bytes pointed to by the faulting RIP. It is an
information disclosure of the host kernel stack: the out-of-bounds data
has a direct influence on the instruction emulation and it doesn't seem
complicated to leak host kernel stack data by observing the side effects
of the altered emulation.
Typically, if the destination of a MOVQ is encoded in 8 bytes located
outside of the 15-byte stack buffer, the MOVQ will be applied to a
destination address that is encoded by host kernel stack data. By
scanning its pages to see at which address data got written, the guest
can trivially infer what were the values of the 8 bytes of host kernel
stack.

Signed-off-by: Wenchao Wang <wenchao.wang@intel.com>