cpuid: Enable all CPUIDs to be configurable.
Open all supported CPUID leaves for user space as configurable.
Implement the set_leaf() callback functions for all supported CPUID
leaves in the CPUID controller. These setting functions are mainly used
for the validity processing of the setting values. When the setting
function is NULL, the input value will be completely accepted.

Signed-off-by: Wenchao Wang <wenchao.wang@intel.com>