Calculate XCR0 valid bits
CPUID.(EAX=0DH,ECX=0) is responsible for reporting the settable XCR0
feature bits (see Intel SDM Vol. 2A 3.2 CPUID). Add the valid_xcr0
variable to store this CPUID result of guest. As all VCPUs in the same
VM share the same CPUID in HAXM, this variable is placed in the space
of VM.

* Implement a function that calculates the XCR0 feature bits that can
  be set by guest
* After initializing CPUID and setting CPUID by user, calculate the
  value of valid_xcr0

Signed-off-by: Wenchao Wang <wenchao.wang@intel.com>