Synchronize guest XCR0
Reading and writing of XCR is performed by XGETBV and XSETBV
instructions. By intercepting VM exit caused by the guest executing
XSETBV instruction, the XCR value to be set by the guest can be
recorded. Currently HAXM only supports XCR0. It ensures that XCR0 of
the guest and XCR0 of the VCPU in HAXM are synchronized, and the value
of XCR0 can be loaded correctly during VM entry.

* Implement the interception of the VM exit caused by the guest
  executing XSETBV instruction
* Implement functions to set and get VCPU XCR

Signed-off-by: Wenchao Wang <wenchao.wang@intel.com>