installer: Improve user interaction
Improve the processing flow according to the return value of check
tool: prohibits user to install HAXM only when the system hardware or
OS does not support; allows user to retry when VMs are occupying HAXM;
and if the system needs to be configured, prompts a warning but proceed
installation. When uninstalling, only the occupancy of VMs is checked.

* Add status and flags according to the error code of check tool
* Update the function of checking environment to comply with check tool
  upgrade
* Add message boxes to show and handle the check result for both
  installation and uninstallation process

Signed-off-by: Wenchao Wang <wenchao.wang@intel.com>