Ensure the external interface unchanged
Since `struct vcpu_state_t` is one of the parameters of the external
interfaces, defining the variable of the extended control register XCR0
in it will cause the change of the external interface. Considering the
compatibility with previous versions, move the definition of XCR0 to
`struct vcpu_t` to ensure the external interface remains unchanged.

* Removed the definition of XCR0 from the external interface
* Adjusted corresponding variables access, function parameter
  definitions and invocations

Signed-off-by: Wenchao Wang <wenchao.wang@intel.com>